Move Connection and Cursor protocol definition different module
